Remove unused passkey auth, add preload
I removed the unused passkey authentication and associated settings and
templates. I also added the htmx preload extention and cache headers.
This required the use of django's cache control helper as well as the
vary on headers helper to cache different versions for hx-request and
non-hx request so the partials and full pages don't interfere. Django's
cache page does not respect this, so using that seemed to break things.

I also brought down the current version of alpine to serve directly
instead of from the cdn.

Signed-off-by: Rick Henry <rickhenry@rickhenry.dev>